==> studienarbeit/party_programs/03_plotting.ipynb <==
import pandas as pd
import seaborn as sns
import json

from studienarbeit.utils.plots import Plots
from studienarbeit.utils.load import EDataTypes
from pathlib import Path
from dotenv import load_dotenv
from nltk.tokenize import sent_tokenize
from tqdm import tqdm

load_dotenv()
sns.set(style="white", palette="muted", rc={"figure.figsize": (20, 8)})
tqdm.pandas()

data_dir = Path("../../data/party_programs")
data_type = EDataTypes.PARTY_PROGRAMS

with open("../../data/party_colors.json", "r", encoding="utf-8") as f:
    party_palette = json.load(f)
plots = Plots(data_type, data_dir, party_palette)

### Standard-Datensatz

df_party_programs = pd.read_parquet(data_dir / "party_programs.parquet")
df_party_programs

with open("../../data/party_encoding.json", "r", encoding="utf-8") as f:
    party_encoding = json.load(f)
party_encoding = {v: k for k, v in party_encoding.items()}
df_party_programs["party"] = df_party_programs["party"].map(party_encoding)
df_party_programs

plots.party_count(df_party_programs)

df_party_programs["char_count"] = df_party_programs["clean_text"].progress_apply(len).astype("int16")
df_party_programs["word_count"] = (
    df_party_programs["clean_text"].progress_apply(lambda x: len(x.split())).astype("int16")
)
df_party_programs["sentence_count"] = (
    df_party_programs["clean_text"].progress_apply(lambda x: len(sent_tokenize(x))).astype("int16")
)

plots.word_count(df_party_programs, "word_count", "Anzahl an Wörtern nach Partei", 350, "Anzahl an Wörtern")

### Datensatz mit Sätzen als Einträgen

df_party_programs_sentence = pd.read_parquet(data_dir / "party_programs_sentence.parquet")
df_party_programs_sentence

df_party_programs_sentence["party"] = df_party_programs_sentence["party"].map(party_encoding)
df_party_programs_sentence

plots.party_count(df_party_programs_sentence)

df_party_programs_sentence["char_count"] = df_party_programs_sentence["clean_text"].progress_apply(len).astype("int16")
df_party_programs_sentence["word_count"] = (
    df_party_programs_sentence["clean_text"].progress_apply(lambda x: len(x.split())).astype("int16")
)
df_party_programs_sentence["sentence_count"] = (
    df_party_programs_sentence["clean_text"].progress_apply(lambda x: len(sent_tokenize(x))).astype("int16")
)

plots.word_count(df_party_programs_sentence, "word_count", "Anzahl an Wörtern nach Partei", 50, "Anzahl an Wörtern")

### Wordclouds (allgemein)

plots.wordclouds(df_party_programs)
